Updated the files
I ran into an issue with the example for hierarchical models: It seems like in some cases the covariance matrix of bmhat is rank deficient, and cholesky factorization as it is coded now does not work. It should still be possible to cholesky factorize into some square root, which is all I need. I was figuring out how the function chol() works in 'QR decomposition' file, but I haven't resolved the issue...