The test for GLMMs
This is the implementation of the test that is closest to the LMM case, I proceed in the same way, except that X becomes sqrtW*X,Z becomes sqrtW*Z, and we obtain bhat=P*sqrtW*z instead of Py. z is not normal, but its covariance matrix can be calculated (its the same as in PQL), and then we obtain the covariance for bhat. (and assume normality). The covariance is known under H1 and under H0, all I change for H0 is psi, the covariance of b. W I suppose to stay the same (perhaps this is a problem?). The results seem weird, although at least, when I  put in the covariance under H1, not H0, the p-values do seem to uniformly range from 0-1, as one would expect (I plotted 100 tests, the p-values are approximately a straight line from 0 to 1).