The test for GLMMs, updated version
I added a second test, glmmPQL, that takes a different route and uses a PQL-style argument to determine the normal distribution of f (note that the f in the book is not the same as the f in the test, so I had to adapt the proof/reasoning). This test seems to work, and seems to avoid some of the numerical problems (Cholesky decomposition) of the first approach. Still not sure why the first test seems to perform weirdly...